Refactor en metodos que retornan usuario en UsuarioService

+++ src/main/java/org/alvarub/workouttrackerproject/persistence/dto/usuario/UsuarioResponseDTO.java
@@ -24,7 +24,7 @@ public class UsuarioResponseDTO {
 
     private LocalDateTime updatedAt;
 
-    private Long bodyWeight;
+    private Double bodyWeight;
 
     private Long completedWorkouts;
